requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
mongoengine>=0.27.0
pymongo[srv]==3.12
celery>=5.3
redis>=5.0
numpy>=1.26
pandas>=2.1
matplotlib>=3.8
pdfplumber>=0.10
python-docx>=1.1
pytesseract>=0.3.10
Pillow>=10.0
reportlab>=4.0
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.2
whitenoise>=6.6
uvicorn>=0.23
mdutils>=1.6
django-compressor>=4.4
django-libsass>=0.9
